Address feedback after code review
